Update Format & lint.md
I propose improving the English on the site so that it reflects proper grammar, spelling, punctuation, and voice. In all proposed changes I will make, my goal is not to edit the content as little as possible and only correct obvious errors. 

A long-term goal that you should consider is making sure the site copy is written in the same voice: third person. Right now it is written in a mix of first-person singular and others. I'm not changing that now, but I propose it as something to consider as a long-term goal we can work on.